flash.py: Refactor so that flash.Flash is not a Device subclass.
It's not really necessary for a flash memory to be implemented as a
Device so long as you can read/write to it.  By splitting this up, we
can also do tricks like initialize the Device part first so that we can
then probe the registers to get things like flash sector size or total
size which we then pass in to the Flash initializer.