probe.py: Halt the target CPUs before trying to probe past the ROM tables.
There are plenty of reasons why it's probably a bad idea to probe a
running target.  The reason we chose here is that if an STM is in Sleep
mode (by executing a WFI instruction), it seems that we can't properly
read some of the Flash registers.  Probably something in the RCC is set
to disable them while the target is in Sleep mode.  In any case, halting
the target brings it out of Sleep mode and if the target is properly
configured to allow clocking while the debugger is attached then we'll
be able to further probe its devices, including its flash.