fus_tool.py: Add option to set back to booting from flash.
Performing FUS operations requires us to switch to booting from SRAM1 so
that MCU resets don't cause whatever firmware is in flash to
inadvertently mess up our transactions with the CPU2 firmware.  It's
convenient to have a command to switch us back to booting from flash
when we are done with all our FUS operations.