ipc/system_channel.py: Add method to tell the BLE firmware to execute.
This is different from "starting" the firmware.  The firmware starts
when you execute ipc.start_firmware() and eventually generates an event
telling us that it is ready to accept commands.  The firmware is then in
a kind of pre-initialized state where you then need to execute a
firmware-specific initialization command (BLE_INIT for the BLE stack) to
tell it what operational parameters it should use.  It's also essential
that the system clocks all be configured for proper operation before
executing that initialization command.