psdb/device.py: Replace Device's optional target field with a required owner field.
Devices used to be associated with a specific target, where a target was
something like an instance of a SoC.  However, in multiprocessor
systems, there can be multiple copies of the same device when they are
Cortex-defined per-CPU devices (that have to be accessed from each CPU's
dedicated AP).  It doesn't make sense for those devices to be associated
with a target; instead they should be associated with a particular CPU.

This commit changes the Device class so that it no longer records a
"target" at all and instead records the concept of an owner.  All
Devices now have to be owned by something, but it doesn't have to be a
target.  This does necessitate manually recording a target field down in
some subclasses where they actually had to make use of the target as
part of their normal operaionts (mainly flash devices which make use of
the target to optimize the debug probe clock frequency, for instance).