cpus/cortex.py: Refactor many cpu-control methods out into scs_base.py.
This also changes our management of the DHCSR.C_DEBUGEN bit -
previously, we would turn it on and off depending on if the CPU was
meant to be running or not.  However, if we want to support breakpoints
then this bit must always be enabled.  We now set it once at init time
and keep it set thereafter.