gdb_tool.py: Track target state so we can handle it stopping autonomously.
A target that has a breakpoint set (or perhaps just a BKPT instruction)
will be able to stop itself without our intervention.  This commit
tracks whether or not we think the target is halted or running, and if
it is running instead of waiting forever for a gdb command we will
periodically check the target to see if it has halted itself.