gdb_tool.py: Add support for using the hardware breakpoint unit.
This allows us to add breakpoints to the system.  For now, we use
hardware breakpoints everywhere (even if gdb requests a software one).
In the future it would be helpful to insert breakpoint instruction for
any code that is running from RAM, although that is a very low priority
since embedded code will almost always be running from flash.