stm32g0/flash.py: Refactor to use common stm32.flash_type1 base.
This is the exact same flash device used in many G4 variants, except
that in the documentation the BSY bit has been renamed BSY1 for some
reason.  We rename the bit to BSY here so that the base class will be
able to poll it properly, and accept the divergence from the official
reference manual.