psdb/probes: Defer probe enumeration until someone actually tries to access a probe.
Previously, we would enumerate all probes when importing psdb.probes.
However, for clients that import psdb.probes but then don't actually use
it at runtime (say, because they are trying to perform an operation on a
target over either SWD or some other in-band protocol that is
user-selectable) this can cause issues, especially on Linux if USB
access to the probes isn't exposed to unprivileged users.