devices/device.py: Reserve all capitalized names with a leading underscore for register names and detect faulty accesses.
Previously, if you tried to assign to a non-existent register this would
just add a new field to the Device object with an integer value rather
than writing anything to a device.  This could lead to subtle bugs where
you have a typo in a register name somewhere.

On the Device object we now assume that all attrs of the form _FOO are
register names and that an attempt to write to such an attr that doesn't
exist in the register dict is erroneous.