psdb/gdb_tool.py: Allow specifying which CPU in a multi-core MCU to target.
The STM32H7 series has multiple cores in a single MCU.  Each core may
run its own separate binary compiled using M4 or M7 options.  In that
case, we'd like to target which CPU to debug against.

We could also have the case where both CPUs are running the same binary
and then we'd like to treat each CPU as though it were a separate
thread.  Currently we don't support that style of debugging.