stm32g4/stm32g4.py: Provide a pre-probe method to configure the DBGMCU.
This is similar to what we do with the H7 series, however the G4 has the
drawback that it only has a single AP that goes through the Cortex core
while the H7 has a dedicated debugger AP.  This means that on the H7 we
can enable the DBGMCU clocks even if both H7 cores are in low-power
mode, while on the G4 we can't access the DBGMCU from low-power mode
since the single AP going through the G4 core will be stopped.

However, this does allow us to connect under reset, automatically
configure the DBGMCU and then release the core to run normally.  If we
boot the G4 core in this way, we will then later be able to connect to
it even if it is in low-power mode without requiring the G4 firmware to
have enabled anything in the DBGMCU on our behalf.  For instance, the
following command is sufficient to boot the G4 in a mode where we could
later debug it from a low-power state:

    psdb_flash_tool --connect-under-reset

This simply connects to the G4 while asserting SRST - during the probing
process the STM32G4.pre_probe() method will configure DBGMCU, then the
whole flash operation will be a no-op since no file was provided and
finally before exiting the MCU will be released from SRST at which point
it will boot with DBGMCU configured for debugging in low-power modes.