stlink/stlink.py: Remove _should_offload_ap() tests.
When this code was originally written, it had seemed that
CSW.DbgSwEnable would get clobbered by the ST firmware when we tried to
access APB APs.  The only two boards that I have ever had that have APB
APs are a Nucleo-H743ZI2 and a Nucleo-H745ZI-Q.  Both of these boards
have V3J2 firmware and neither of them seems to have any issues with
CSW.DbgSwEnable clobbering with the current code base.  It's possible
that adding support for OpenAP() may have resolved it, but even by
removing that code I haven't been able to reproduce any clobbering, so
it may have been some other bug in psdb that was clobbering CSW rather
than the ST firmware.

In any case, I'm removing this nastiness since it seems completely
unnecessary now.