probes/probe.py: Factor deassertion of SRST out of XDS110 code into generic code.
The TI MCUs don't support probing while SRST is asserted, so when the
original XDS110 driver was written we had a line of code to deassert
SRST in case it had been left asserted from a previous run.  That's
actually undesirable, especially if we are using an XDS110 to debug an
STM32 MCU which does support connecting under reset because then it will
take the STM32 out of reset as part of the debugger connect process.