psdb/access_port.py: Refactor so that detecting a raw AP doesn't raise an Exception.
The base AP class doesn't take a csw_base argument to its initializer,
but the IDRMapper() factory code was passing one in.  If for some reason
we detected a raw AP instead of a MemAP, this would trigger an exception
due to the mismatched argument list.

With a bit of refactoring, we can avoid the explicit argument list and
support both raw APs and specialized APs.