stlink/stlink.py: If the probe is already in debug mode, don't disconnect and reconnect.
A disconnect (LeaveDebugMode) will deassert SRST.  If the firmware on
the target then boots up and immediately enters a low-power state, a
subsequent SWDConnect call will fail because we won't be able to halt
the CPU out of reset.

So, if we are already in Debug mode, don't leave it before doing our
regular SWDConnect.  We can still issue SWDConnect if we've already
issued it and it will do the JTAG -> SWD sequence to connect to the
target which is what we need if the connection was invalidated due to a
target reset.  This allows us to maintain SRST for the full operation.