stm32g4/stm32g4.py: Synchronize/desynchronize FB_MODE and BFB2 settings when resetting and resuming.
ST is really annoying.  The BFB2 setting is stored in options flash and
when BFB2=1 is set the MCU boots through the bootloader instead of
through the reset handler.  When we use VC_CORERESET to catch a reset,
if BFB2=1 we catch it upon entry to the bootloader.  We'll find in that
case that BFB2=1 but FB_MODE=0 so nothing is swapped.  This makes things
like inspect_tool show the wrong flash banks and flash_tool just... do
the wrong thing.

To fix this, we can set FB_MODE=BFB2 and then flash_tool and
inspect_tool do the right thing.  However, this triggers a second
problem with the ST firmware: if we then resume the processor and allow
the bootloader to run, it will then *toggle* FB_MODE (presumably it
thinks it is toggling it from 0 to 1, but since we already set it to 1
it toggles it from 1 to 0) and then try to run from the wrong bank.

So, the whole fix is to remember the setting of FB_MODE when we come out
of reset, set FB_MODE=BFB2 while the CPU is halted and we are doing
debugging and finally set FB_MODE back to its original setting before
resuming the CPU.

With all that in place, bank swapping while connecting under reset
mostly works properly... except for a small bug in the flash driver that
we fix in the next commit.