stm32g4/flash_3.py: Compute bker properly, especially out of reset.
When we connect under reset and catch the CPU in the reset handler, the
SYSCFG device hasn't even been enabled yet and the flash driver was
making no attempt to do so.  Coupled with the FB_MODE weirdness dealt
with in the previous commit, we can drastically simplify setting the
BKER bit to the correct value.

With this change, flashing, erasing and bank swapping all work properly
even when connecting under reset.