component/component.py: Probe using DFS instead of BFS.
For some unknown reason we were searching using breadth-first search.
This resulted in printing the component hierarchy in the wrong order,
sometimes implying that components were children of components that they
were not.

Depth-first search seems to work fine on the handful of boards I've
tested, so switch to that for now.