stm32u5: Add flash driver.
This is heavily based on the STM32G4 flash driver, however unlike the
STM32G4, the STM32U5 behaves in a sane manner when using the
bank-swapping feature.  The STM32U5 hardware swaps the banks before
making the first VTOR fetch, meaning that even if we connect under reset
vector catch the banks will already be swapped the way they should be.