psdb/probes: Refactor enumeration so that it doesn't require device allocation.
When multiple probes are connected to the system at once, previously
psdb would have to claim all probes just to use one of them.  Once a
probe was claimed, it no longer is available for another process to use.
This is quite annoying when trying to use multiple probes on a single
system independently.

By changing the enumeration to only look at things like vendor or
product IDs that don't require device access we can solve the issue.