psdb: Change flash max_write_speed property to max_nowait_write_speed and require correct probe behaviour.
The flash peripherals had "maximum" write speeds because when we tried
to write to them with the STLINK above those speeds the STLINK would
error out.  However, we are able to write to them with the XTSWD above
those speeds just fine; it seems that the problem is actually that the
STLINK does not handle WAIT states properly when the flash stalls the
bus.

The most recent STLINK firmware does seem to address this for the most
part, allowing writes at 24 MHz in most cases.  Performance does not
change since the write performance is limited by the flash response time
and not the SWD frequency, but at least it no longer errors out...
except in the case of U5 where it now works at 8 MHz but still errors
out at 24 MHz.

Anyways, by providing the max no-WAIT write frequency on the flash we
can then make it the debug probe's choice what to do rather than
providing a strict max write frequency.