psdb: Refactor how TCK frequencies are selected.
Previously, we had the concept of the max frequency at which a debug
probe could operate.  However, that makes little sense since we actually
have the following limits:

1. The maximum frequency that the target supports for reading or writing.
2. The maximum frequency that the probe supports for reading or writing.
3. The maximum frequency that the hardware configuration (cabling)
   supports.

We change the APIs so that we select the maximum target frequency rather
than the maximum probe frequency, and rely on the probe implementation
limiting to frequency to the maximum supported by the probe.  For
certain versions of the ST-LINK firmware we also add checks to ensure
that we don't exceed a frequency at which the target may give back WAIT
acknowledgements over SWD since those are not handled well by some
versions of the ST-LINK firmware.

We will address a user-configurable global limit due to poor cabling
situations in some future commit.