stm32g0/stm32g0.py: Refactor probing to be more like other MCUs.
The G0 is a bit weird in that when it is held in reset you can't do
things like read the DBGMCU IDCODE properly.  So, we rearrange how
things are done a bit here.  It also seems like the writes to enable the
DBGMCU clocks don't do anything but they are kept there for now.